% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_R6_elements.R
\name{actionButton}
\alias{actionButton}
\title{R6 Class employing shiny::actionButton}
\format{
An \link[R6:R6Class]{R6::R6Class} object.
}
\description{
An instance of this class is used in similar way to shiny::actionButton
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{actionButton$new()}}
\item \href{#method-ui_input}{\code{actionButton$ui_input()}}
\item \href{#method-ui_output}{\code{actionButton$ui_output()}}
\item \href{#method-ui_render_output}{\code{actionButton$ui_render_output()}}
\item \href{#method-server}{\code{actionButton$server()}}
\item \href{#method-get_uiInId}{\code{actionButton$get_uiInId()}}
\item \href{#method-get_uiOtId}{\code{actionButton$get_uiOtId()}}
\item \href{#method-get_uiRdOtId}{\code{actionButton$get_uiRdOtId()}}
\item \href{#method-clone}{\code{actionButton$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{actionButton$new(.label_)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ui_input"></a>}}
\if{latex}{\out{\hypertarget{method-ui_input}{}}}
\subsection{Method \code{ui_input()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{actionButton$ui_input(
  .tag_ = "UI_input",
  .class_ = "ml-2 pt-2 d-flex align-items-center\\n                        text-right",
  .style_ = "display: flex;\\n                        margin-top: 2rem !important;",
  .width_ = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ui_output"></a>}}
\if{latex}{\out{\hypertarget{method-ui_output}{}}}
\subsection{Method \code{ui_output()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{actionButton$ui_output(.tag_ = "UI_output")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ui_render_output"></a>}}
\if{latex}{\out{\hypertarget{method-ui_render_output}{}}}
\subsection{Method \code{ui_render_output()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{actionButton$ui_render_output(.tag_ = "UI_render_output")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-server"></a>}}
\if{latex}{\out{\hypertarget{method-server}{}}}
\subsection{Method \code{server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{actionButton$server(input, output, session)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_uiInId"></a>}}
\if{latex}{\out{\hypertarget{method-get_uiInId}{}}}
\subsection{Method \code{get_uiInId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{actionButton$get_uiInId()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_uiOtId"></a>}}
\if{latex}{\out{\hypertarget{method-get_uiOtId}{}}}
\subsection{Method \code{get_uiOtId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{actionButton$get_uiOtId()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_uiRdOtId"></a>}}
\if{latex}{\out{\hypertarget{method-get_uiRdOtId}{}}}
\subsection{Method \code{get_uiRdOtId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{actionButton$get_uiRdOtId()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{actionButton$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
