% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs_.R
\name{check_missings_}
\alias{check_missings_}
\title{Check if data structure contain missing (NA) values.}
\usage{
check_missings_(.data_, .label_)
}
\arguments{
\item{.data_}{The data structure to check for missing values.}

\item{.label_}{The name of the data structure being checked.}
}
\value{
A logical or message based on whether missing data exists.
}
\description{
Check if data structure contain missing (NA) values.
}
\examples{
\dontrun{
}
}
