% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs_.R
\name{check_numerics_}
\alias{check_numerics_}
\title{Check if data structure columns are all numerics.}
\usage{
check_numerics_(.data_, .label_)
}
\arguments{
\item{.data_}{The data structure to check if all are numerics.}

\item{.label_}{The name of the data structure being checked.}
}
\value{
A logical or message based on whether non-numeric columns exists.
}
\description{
Check if data structure columns are all numerics.
}
\examples{
\dontrun{
}
}
