% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs_.R
\name{check_options_}
\alias{check_options_}
\title{Check if two or more decision options data were provided.}
\usage{
check_options_(.data_, .label_)
}
\arguments{
\item{.data_}{The data structure to check for number of decision options.}

\item{.label_}{The name of the data structure being checked.}
}
\value{
A logical or message based on whether missing data exists.
}
\description{
Check if two or more decision options data were provided.
}
\examples{
\dontrun{
}
}
