% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs_.R
\name{check_PSA_inputs}
\alias{check_PSA_inputs}
\title{Check PSA data structures passed to a shiny app}
\usage{
check_PSA_inputs(.costs_, .effs_, .params_, .id_)
}
\arguments{
\item{.costs_}{A tibble/dataframe/matrix containing costs data.}

\item{.effs_}{A tibble/dataframe/matrix containing effects data.}

\item{.params_}{A tibble/dataframe/matrix containing parameters data.}

\item{.id_}{A letter "c", "e" or "p" that declares if the function is
to check costs, effects or parameters data structures.}
}
\description{
Check PSA data structures passed to a shiny app
}
\examples{
\dontrun{
}
}
