% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stability.R
\name{check_PSA_stability}
\alias{check_PSA_stability}
\title{Check stability of PSA outputs
This function produces a set of plots to allow modellers to investigate if
the number of PSA runs were sufficient or not!}
\usage{
check_PSA_stability(
  .PSA_data = NULL,
  .effs = NULL,
  .costs = NULL,
  .interventions = NULL,
  ...
)
}
\arguments{
\item{.PSA_data}{A list of class shinyPSA that contains summary PSA
results.}

\item{.effs}{A tibble containing the \code{effects} from PSA. Number of
\code{columns} is equal to the interventions while the number of
\code{rows} is equal to the number of PSA simulations to be summarised.}

\item{.costs}{A tibble containing the \code{costs} from PSA. Number of
\code{columns} is equal to the interventions while the number of
\code{rows} is equal to the number of PSA simulations to be summarised.}

\item{.interventions}{A vector containing the names of all interventions.
If not provided or less names than needed is provided, the function will
generate generic names, for example \code{intervention 1}.}

\item{...}{Additoinal plot arguments, which currently supports:
.legend_pos (legend position): format (x, y) where x and y are between
\code{0, 1}. If missing, the function uses default (0.8, 0.85).}
}
\value{
A list with objects of class ggplot.
}
\description{
Check stability of PSA outputs
This function produces a set of plots to allow modellers to investigate if
the number of PSA runs were sufficient or not!
}
\examples{
\dontrun{
library(ShinyPSA)

p <- check_PSA_stability(
         .effs = ShinyPSA::Hypothetical_PSA$e,
         .costs = ShinyPSA::Hypothetical_PSA$c,
         .interventions = ShinyPSA::Hypothetical_PSA$treats)

p[["Effects stability"]]
p[["Costs stability"]]
p[["Cost per effect stability"]]
}
}
