% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs_.R
\name{check_uniqueness_}
\alias{check_uniqueness_}
\title{Check if enough unique data points are in supplied parameters object.}
\usage{
check_uniqueness_(.data_, .label_)
}
\arguments{
\item{.data_}{The parameters data structure to check for missing values.}

\item{.label_}{The name of the data structure being checked.}
}
\value{
A logical or message based on whether missing data exists.
}
\description{
Check if enough unique data points are in supplied parameters object.
}
\examples{
\dontrun{
}
}
