% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs_.R
\name{coerce_numerics_}
\alias{coerce_numerics_}
\title{Remove formatting from data and ensure all columns are numericals.}
\usage{
coerce_numerics_(.data_)
}
\arguments{
\item{.data_}{The data object from which to remove formatting and
convert all columns to numeric.}
}
\description{
Remove formatting from data and ensure all columns are numericals.
}
\examples{
\dontrun{
numerical_parameters <- coerce_numerics_(
   .data_ = ShinyPSA::Clopidogrel_PSA$p)
}
}
