% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_NMBs_.R
\name{compute_CEACs_}
\alias{compute_CEACs_}
\title{Compute Cost-Effectiveness Acceptability Curve (CEAC)}
\usage{
compute_CEACs_(
  .nmb,
  .effs = NULL,
  .costs = NULL,
  .interventions = NULL,
  .Kmax = NULL,
  .wtp = NULL
)
}
\arguments{
\item{.nmb}{A list (with similar features to a 3D-array) containing the
Net Monetary Benefits from each probabilistic sensitivity analysis (PSA)
run for each intervention across a range of willingness-to-pay (WTP)
values. The dimensions of this list are:
\code{List:WTP, Tibble(Rows: PSA simulations, Cols: Interventions)}.}

\item{.effs}{A tibble containing the \code{effects} from PSA. Number of
\code{columns} is equal to the interventions while the number of
\code{rows} is equal to the number of PSA simulations to be summarised.}

\item{.costs}{A tibble containing the \code{costs} from PSA. Number of
\code{columns} is equal to the interventions while the number of
\code{rows} is equal to the number of PSA simulations to be summarised.}

\item{.interventions}{A vector containing the names of all
interventions. If not provided or less names than needed is provided,
the function will generate generic names, for example
\code{intervention 1}.}

\item{.Kmax}{The maximum willingness-to-pay threshold to use in the
analysis. This parameter is ignored if \code{wtp} is provided.}

\item{.wtp}{A vector of numerical values declaring the
willingness-to-pay (WTP) values to use in the analysis. If \code{NULL}
(default) a range of WTP values (up to \code{.Kmax} will be used.}
}
\value{
A tibble containing the probability of being cost-effective
for all interventions.
}
\description{
Compute Cost-Effectiveness Acceptability Curve (CEAC)
}
\examples{
\dontrun{}
}
