% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_NMBs_.R
\name{compute_CEAFs_}
\alias{compute_CEAFs_}
\title{Compute Cost-Effectiveness Acceptability Frontier}
\usage{
compute_CEAFs_(.ceac, .nmb = NULL)
}
\arguments{
\item{.ceac}{A tibble containing the probability of being cost-effective
for all interventions.}

\item{.nmb}{A list (with similar features to a 3D-array) containing the
Net Monetary Benefits from each probabilistic sensitivity analysis (PSA)
run for each intervention across a range of willingness-to-pay (WTP)
values. The dimensions of this list are:
\code{List:WTP, Tibble(Rows: PSA simulations, Cols: Interventions)}.}
}
\value{
A tibble containing the probability of being cost-effective
for all interventions alongside the CEAF.
}
\description{
Compute Cost-Effectiveness Acceptability Frontier
}
\examples{
\dontrun{}
}
