% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_EVPPIs_.R
\name{compute_EVPPIs_}
\alias{compute_EVPPIs_}
\title{Compute the Expected Value of Perfect Partial Information (EVPPI).}
\usage{
compute_EVPPIs_(
  .PSA_data,
  .effs = NULL,
  .costs = NULL,
  .params = NULL,
  .set = NULL,
  .set_names = NULL,
  .subset_ = FALSE,
  .MAICER_ = 30000,
  .units_ = "£",
  .individual_evppi_ = TRUE,
  .discount_rate_ = 0.035,
  .evppi_population_ = NULL,
  .time_horion_ = NULL,
  .session = NULL
)
}
\arguments{
\item{.PSA_data}{A list of class shinyPSA that contains summary PSA
results.}

\item{.effs}{A matrix containing the \code{effects} from PSA. Number of
\code{columns} is equal to the interventions while the number of
\code{rows} is equal to the number of PSA simulations to be summarised.}

\item{.costs}{A matrix containing the \code{costs} from PSA. Number of
\code{columns} is equal to the interventions while the number of
\code{rows} is equal to the number of PSA simulations to be summarised.}

\item{.params}{A matrix containing parameters' configurations used in
the PSA.}

\item{.set}{A vector of parameters' names for conditional EVPPI.}

\item{.set_names}{A vector of parameter-names to be used for EVPPI.}

\item{.subset_}{Boolean for whether to estimate conditional EVPPI for a
subset of parameters.}

\item{.MAICER_}{The Maximum acceptable incremental cost-effectiveness ratio.
(MAICER) to be considered in the summary table. Default value is
\code{30,000}.}

\item{.units_}{A character, the units to associate with the
monitory values in the summary table. Default is sterling pounds
(GBP) \code{"\u00A3"}.}

\item{.individual_evppi_}{Logical (default \code{TRUE}) to return per person
EVPPI, otherwise population EVPPI will be reported.}

\item{.discount_rate_}{The discount rate used to discount future affected
populations.}

\item{.evppi_population_}{The size of the population that is annually
affected by the competing health technologies under evaluation.}

\item{.time_horion_}{The time expected to pass (in years) before the
interventions under consideration change (how long before the decision
under consideration become obsolete or requires updating).}

\item{.session}{Shiny app session.}
}
\value{
A list containing the EVPPI results table and caption information.
}
\description{
Compute the Expected Value of Perfect Partial Information (EVPPI).
}
\examples{
\dontrun{
library(ShinyPSA)

}
}
