% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ICERs_.R
\name{compute_ICERs_}
\alias{compute_ICERs_}
\title{Compute ICER(s)}
\usage{
compute_ICERs_(.icer_data, .effs = NULL, .costs = NULL, .interventions = NULL)
}
\arguments{
\item{.icer_data}{A table containing average costs and QALYs data}

\item{.effs}{A tibble containing the \code{effects} from PSA. Number of
\code{columns} is equal to the interventions while the number of
\code{rows} is equal to the number of PSA simulations to be summarised.}

\item{.costs}{A tibble containing the \code{costs} from PSA. Number of
\code{columns} is equal to the interventions while the number of
\code{rows} is equal to the number of PSA simulations to be summarised.}

\item{.interventions}{A vector containing the names of all
interventions. If not provided or less names than needed is provided,
the function will generate generic names, for example
\code{intervention 1}.}
}
\value{
A dataframe with data from icer_data in addition to
\code{qalys and costs diffrential(s)}, \code{dominance} & \code{icer(s)}
}
\description{
Compute ICER(s)
}
\examples{
\dontrun{}
}
