% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_NMBs_.R
\name{compute_NMBs_}
\alias{compute_NMBs_}
\title{Compute Monetary Net-Benefit (NMB) or incremental NMB (iNMB)}
\usage{
compute_NMBs_(.effs, .costs, .interventions = NULL, .Kmax = NULL, .wtp = NULL)
}
\arguments{
\item{.effs}{A tibble containing the \code{effects} from PSA. Number of
\code{columns} is equal to the interventions while the number of
\code{rows} is equal to the number of PSA simulations to be summarised.}

\item{.costs}{A tibble containing the \code{costs} from PSA. Number of
\code{columns} is equal to the interventions while the number of
\code{rows} is equal to the number of PSA simulations to be summarised.}

\item{.interventions}{A vector containing the names of all
interventions. If not provided or less names than needed is provided,
the function will generate generic names, for example
\code{intervention 1}.}

\item{.Kmax}{The maximum willingness-to-pay threshold to use in the
analysis. This parameter is ignored if \code{wtp} is provided.}

\item{.wtp}{A vector of numerical values declaring the
willingness-to-pay (WTP) values to use in the analysis. If \code{NULL}
(default) a range of WTP values (up to \code{.Kmax} will be used.}
}
\value{
A list containing the NMB list, eNMB tibble, WTP tibble and
other objects.
}
\description{
Compute Monetary Net-Benefit (NMB) or incremental NMB (iNMB)
}
\examples{
\dontrun{}
}
