% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_R6_elements.R
\name{dataTableGT}
\alias{dataTableGT}
\title{R6 Class employing gt::render_gt & gt::gt_output}
\format{
An \link[R6:R6Class]{R6::R6Class} object.
}
\description{
An instance of this class is used in similar way to gt::render_gt &
gt::gt_output
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-dataTableGT-new}{\code{dataTableGT$new()}}
\item \href{#method-dataTableGT-ui_input}{\code{dataTableGT$ui_input()}}
\item \href{#method-dataTableGT-ui_output}{\code{dataTableGT$ui_output()}}
\item \href{#method-dataTableGT-ui_render_output}{\code{dataTableGT$ui_render_output()}}
\item \href{#method-dataTableGT-server}{\code{dataTableGT$server()}}
\item \href{#method-dataTableGT-get_uiInId}{\code{dataTableGT$get_uiInId()}}
\item \href{#method-dataTableGT-get_uiOtId}{\code{dataTableGT$get_uiOtId()}}
\item \href{#method-dataTableGT-get_uiRdOtId}{\code{dataTableGT$get_uiRdOtId()}}
\item \href{#method-dataTableGT-clone}{\code{dataTableGT$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataTableGT-new"></a>}}
\if{latex}{\out{\hypertarget{method-dataTableGT-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataTableGT$new(.label_)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataTableGT-ui_input"></a>}}
\if{latex}{\out{\hypertarget{method-dataTableGT-ui_input}{}}}
\subsection{Method \code{ui_input()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataTableGT$ui_input(.tag_ = "UI_input")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataTableGT-ui_output"></a>}}
\if{latex}{\out{\hypertarget{method-dataTableGT-ui_output}{}}}
\subsection{Method \code{ui_output()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataTableGT$ui_output(.tag_ = "UI_output")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataTableGT-ui_render_output"></a>}}
\if{latex}{\out{\hypertarget{method-dataTableGT-ui_render_output}{}}}
\subsection{Method \code{ui_render_output()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataTableGT$ui_render_output(.tag_ = "UI_render_output")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataTableGT-server"></a>}}
\if{latex}{\out{\hypertarget{method-dataTableGT-server}{}}}
\subsection{Method \code{server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataTableGT$server(
  input,
  output,
  session,
  .table_,
  .width_ = NULL,
  .height_ = NULL,
  .align_ = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataTableGT-get_uiInId"></a>}}
\if{latex}{\out{\hypertarget{method-dataTableGT-get_uiInId}{}}}
\subsection{Method \code{get_uiInId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataTableGT$get_uiInId()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataTableGT-get_uiOtId"></a>}}
\if{latex}{\out{\hypertarget{method-dataTableGT-get_uiOtId}{}}}
\subsection{Method \code{get_uiOtId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataTableGT$get_uiOtId()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataTableGT-get_uiRdOtId"></a>}}
\if{latex}{\out{\hypertarget{method-dataTableGT-get_uiRdOtId}{}}}
\subsection{Method \code{get_uiRdOtId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataTableGT$get_uiRdOtId()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataTableGT-clone"></a>}}
\if{latex}{\out{\hypertarget{method-dataTableGT-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataTableGT$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
