% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{generate_95_ci}
\alias{generate_95_ci}
\title{Generate stats}
\usage{
generate_95_ci(.data_, .interventions, .accuracy_, .units_ = "£")
}
\arguments{
\item{.data_}{A matrix containing the \code{effects, costs, or parameters}
from PSA. Number of \code{columns} is equal to the interventions in the first
two, while the number of \code{rows} is equal to the number of PSA
simulations to be summarised, in all three.}

\item{.interventions}{A vector containing the names of all
interventions. If not provided or less names than needed is provided,
the function will generate generic names, for example
\code{intervention 1}.}

\item{.accuracy_}{Number of digits to round up to.}

\item{.units_}{A character, the units to associate with the
monitory values in the summary table. Default is sterling pounds
(GBP) \code{"\u00A3"}.}
}
\description{
Generate stats
}
