% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_R6_elements.R
\name{inputSelection}
\alias{inputSelection}
\title{R6 Class employing shiny::inputSelection}
\format{
An \link[R6:R6Class]{R6::R6Class} object.
}
\description{
An instance of this class is used in similar way to
shiny::inputSelection
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{inputSelection$new()}}
\item \href{#method-ui_input}{\code{inputSelection$ui_input()}}
\item \href{#method-ui_output}{\code{inputSelection$ui_output()}}
\item \href{#method-ui_render_output}{\code{inputSelection$ui_render_output()}}
\item \href{#method-server}{\code{inputSelection$server()}}
\item \href{#method-get_uiInId}{\code{inputSelection$get_uiInId()}}
\item \href{#method-get_uiOtId}{\code{inputSelection$get_uiOtId()}}
\item \href{#method-get_uiRdOtId}{\code{inputSelection$get_uiRdOtId()}}
\item \href{#method-clone}{\code{inputSelection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{inputSelection$new(.label_)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ui_input"></a>}}
\if{latex}{\out{\hypertarget{method-ui_input}{}}}
\subsection{Method \code{ui_input()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{inputSelection$ui_input(
  .tag_ = "UI_input",
  .choices_,
  .selected_ = NULL,
  .multiple_ = FALSE,
  .width_ = "100\%",
  .class_ = "d-flex align-items-center",
  .style_ = "display: flex; margin-top:\\n                        2rem !important;"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ui_output"></a>}}
\if{latex}{\out{\hypertarget{method-ui_output}{}}}
\subsection{Method \code{ui_output()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{inputSelection$ui_output(.tag_ = "UI_output")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ui_render_output"></a>}}
\if{latex}{\out{\hypertarget{method-ui_render_output}{}}}
\subsection{Method \code{ui_render_output()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{inputSelection$ui_render_output(.tag_ = "UI_render_output")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-server"></a>}}
\if{latex}{\out{\hypertarget{method-server}{}}}
\subsection{Method \code{server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{inputSelection$server(input, output, session, .choices_ = .choices_)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_uiInId"></a>}}
\if{latex}{\out{\hypertarget{method-get_uiInId}{}}}
\subsection{Method \code{get_uiInId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{inputSelection$get_uiInId()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_uiOtId"></a>}}
\if{latex}{\out{\hypertarget{method-get_uiOtId}{}}}
\subsection{Method \code{get_uiOtId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{inputSelection$get_uiOtId()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_uiRdOtId"></a>}}
\if{latex}{\out{\hypertarget{method-get_uiRdOtId}{}}}
\subsection{Method \code{get_uiRdOtId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{inputSelection$get_uiRdOtId()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{inputSelection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
