% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_R6_elements.R
\name{numericInput}
\alias{numericInput}
\title{R6 Class employing shiny::numericInput}
\format{
An \link[R6:R6Class]{R6::R6Class} object.
}
\description{
An instance of this class is used in similar way to shiny::numericInput
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-numericInput-new}{\code{numericInput$new()}}
\item \href{#method-numericInput-ui_input}{\code{numericInput$ui_input()}}
\item \href{#method-numericInput-ui_output}{\code{numericInput$ui_output()}}
\item \href{#method-numericInput-ui_render_output}{\code{numericInput$ui_render_output()}}
\item \href{#method-numericInput-server}{\code{numericInput$server()}}
\item \href{#method-numericInput-get_uiInId}{\code{numericInput$get_uiInId()}}
\item \href{#method-numericInput-get_uiOtId}{\code{numericInput$get_uiOtId()}}
\item \href{#method-numericInput-get_uiRdOtId}{\code{numericInput$get_uiRdOtId()}}
\item \href{#method-numericInput-clone}{\code{numericInput$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-numericInput-new"></a>}}
\if{latex}{\out{\hypertarget{method-numericInput-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{numericInput$new(.label_)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-numericInput-ui_input"></a>}}
\if{latex}{\out{\hypertarget{method-numericInput-ui_input}{}}}
\subsection{Method \code{ui_input()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{numericInput$ui_input(
  .tag_ = "UI_input",
  .width_ = "100\%",
  .class_ = "d-flex align-items-center",
  .value_ = "500, 20000, 30000, 50000",
  .min_ = 0,
  .max_ = 1
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-numericInput-ui_output"></a>}}
\if{latex}{\out{\hypertarget{method-numericInput-ui_output}{}}}
\subsection{Method \code{ui_output()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{numericInput$ui_output(.tag_ = "UI_output")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-numericInput-ui_render_output"></a>}}
\if{latex}{\out{\hypertarget{method-numericInput-ui_render_output}{}}}
\subsection{Method \code{ui_render_output()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{numericInput$ui_render_output(.tag_ = "UI_render_output")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-numericInput-server"></a>}}
\if{latex}{\out{\hypertarget{method-numericInput-server}{}}}
\subsection{Method \code{server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{numericInput$server(input, output, session, .min_, .max_)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-numericInput-get_uiInId"></a>}}
\if{latex}{\out{\hypertarget{method-numericInput-get_uiInId}{}}}
\subsection{Method \code{get_uiInId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{numericInput$get_uiInId()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-numericInput-get_uiOtId"></a>}}
\if{latex}{\out{\hypertarget{method-numericInput-get_uiOtId}{}}}
\subsection{Method \code{get_uiOtId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{numericInput$get_uiOtId()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-numericInput-get_uiRdOtId"></a>}}
\if{latex}{\out{\hypertarget{method-numericInput-get_uiRdOtId}{}}}
\subsection{Method \code{get_uiRdOtId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{numericInput$get_uiRdOtId()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-numericInput-clone"></a>}}
\if{latex}{\out{\hypertarget{method-numericInput-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{numericInput$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
