% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_R6_elements.R
\name{pickerInput}
\alias{pickerInput}
\title{R6 Class employing shinyWidgets::pickerInput}
\format{
An \link[R6:R6Class]{R6::R6Class} object.
}
\description{
An instance of this class is used in similar way to
shinyWidgets::pickerInput
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-pickerInput-new}{\code{pickerInput$new()}}
\item \href{#method-pickerInput-ui_input}{\code{pickerInput$ui_input()}}
\item \href{#method-pickerInput-ui_output}{\code{pickerInput$ui_output()}}
\item \href{#method-pickerInput-ui_render_output}{\code{pickerInput$ui_render_output()}}
\item \href{#method-pickerInput-server}{\code{pickerInput$server()}}
\item \href{#method-pickerInput-get_uiInId}{\code{pickerInput$get_uiInId()}}
\item \href{#method-pickerInput-get_uiOtId}{\code{pickerInput$get_uiOtId()}}
\item \href{#method-pickerInput-get_uiRdOtId}{\code{pickerInput$get_uiRdOtId()}}
\item \href{#method-pickerInput-clone}{\code{pickerInput$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pickerInput-new"></a>}}
\if{latex}{\out{\hypertarget{method-pickerInput-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pickerInput$new(.label_)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pickerInput-ui_input"></a>}}
\if{latex}{\out{\hypertarget{method-pickerInput-ui_input}{}}}
\subsection{Method \code{ui_input()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pickerInput$ui_input(
  .tag_ = "UI_input",
  .choices_,
  .info_,
  .more_info_,
  .class_ = "flex-fill text-left"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pickerInput-ui_output"></a>}}
\if{latex}{\out{\hypertarget{method-pickerInput-ui_output}{}}}
\subsection{Method \code{ui_output()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pickerInput$ui_output(.tag_ = "UI_output")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pickerInput-ui_render_output"></a>}}
\if{latex}{\out{\hypertarget{method-pickerInput-ui_render_output}{}}}
\subsection{Method \code{ui_render_output()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pickerInput$ui_render_output(.tag_ = "UI_render_output")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pickerInput-server"></a>}}
\if{latex}{\out{\hypertarget{method-pickerInput-server}{}}}
\subsection{Method \code{server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pickerInput$server(input, output, session, .choices_ = .choices_)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pickerInput-get_uiInId"></a>}}
\if{latex}{\out{\hypertarget{method-pickerInput-get_uiInId}{}}}
\subsection{Method \code{get_uiInId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pickerInput$get_uiInId()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pickerInput-get_uiOtId"></a>}}
\if{latex}{\out{\hypertarget{method-pickerInput-get_uiOtId}{}}}
\subsection{Method \code{get_uiOtId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pickerInput$get_uiOtId()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pickerInput-get_uiRdOtId"></a>}}
\if{latex}{\out{\hypertarget{method-pickerInput-get_uiRdOtId}{}}}
\subsection{Method \code{get_uiRdOtId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pickerInput$get_uiRdOtId()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pickerInput-clone"></a>}}
\if{latex}{\out{\hypertarget{method-pickerInput-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pickerInput$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
