% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_CEAF_.R
\name{plot_CEAF_}
\alias{plot_CEAF_}
\title{Plot Cost Effectiveness Acceptability Frontier (CEAF)}
\usage{
plot_CEAF_(.PSA_data, ...)
}
\arguments{
\item{.PSA_data}{A list of class shinyPSA that contains summary PSA
results.}

\item{...}{Additional arguments that include:
legend position \code{.legend_pos = c(0.8, 0.85)},
willingness-to-pay threshold(s) \code{.wtp_threshold = c(20000, 30000)},
show WTP threshold(s) lines \code{.show_wtp = TRUE},
show WTP threshold(s) labels \code{.label_wtp = TRUE},
zoom to min/max values \code{.zoom = FALSE},
zoom to supplied coordinates values \code{.zoom_cords = NULL}, and
show 20 points/shapes along the lines \code{.show_shapes = FALSE}.}
}
\value{
An object of class ggplot.
}
\description{
Plot Cost Effectiveness Acceptability Frontier (CEAF)
}
\examples{
\dontrun{
library(ShinyPSA)

PSA_summary <- summarise_PSA_(
  .effs = as_tibble(ShinyPSA::Smoking_PSA$e),
  .costs = as_tibble(ShinyPSA::Smoking_PSA$c),
  .interventions = ShinyPSA::Smoking_PSA$treats)

p <- plot_CEAF_(PSA_summary,
               .legend_pos = NULL,
               .wtp_threshold = c(2000, 10000, 20000, 25000),
               .show_wtp = TRUE,
               .label_wtp = FALSE,
               .zoom = FALSE,
               .zoom_cords = NULL,
               .show_shapes = TRUE,
               .show_title = TRUE)

p
}

}
