% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_CEplane_.R
\name{plot_CEplane_}
\alias{plot_CEplane_}
\title{Plot Cost Effectiveness Plane (CEP).}
\usage{
plot_CEplane_(.PSA_data, ...)
}
\arguments{
\item{.PSA_data}{A list of class shinyPSA that contains summary PSA
results.}

\item{...}{Additional arguments that include:
reference intervention \code{.ref = NULL} rescales interventions on CEP,
legend position \code{.legend_pos = c(0.8, 0.2)},
show ICER information \code{.show_ICER' = TRUE},
nudge ICER labels \code{.nudge_labels' = c(NULL, NULL)},
willingness-to-pay threshold(s) \code{.wtp_threshold = c(20000, 30000)},
show WTP threshold(s) lines and labels \code{.show_wtp = TRUE}, and
seed number to set \code{.seed_no = 1}.}
}
\value{
An object of class ggplot.
}
\description{
Plot Cost Effectiveness Plane (CEP).
}
\examples{
\dontrun{
library(ShinyPSA)

PSA_summary <- summarise_PSA_(
  .effs = as_tibble(ShinyPSA::Vaccine_PSA$e),
  .costs = as_tibble(ShinyPSA::Vaccine_PSA$c),
  .interventions = ShinyPSA::Vaccine_PSA$treats)

p <- plot_CEplane_(PSA_summary,
                 .ref = 1,
                 .show_ICER = TRUE,
                 .legend_pos = c(0.8, 0.2),
                 .show_wtp = FALSE,
                 .zoom = T,
                 .wtp_threshold = c(200),
                 tst = "PRINT", # this will be ignored
                 .nudge_labels = c(0.1, -0.1),
                 .zoom_cords = c(-0.001, 0.001, -5, 5)))

p
}

}
