% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eNMB_.R
\name{plot_eNMB_}
\alias{plot_eNMB_}
\title{Plot Expected Net Monetary Benefit (eNMB)}
\usage{
plot_eNMB_(.PSA_data, ...)
}
\arguments{
\item{.PSA_data}{A list of class shinyPSA that contains summary PSA
results.}

\item{...}{Additional arguments that include:
legend position \code{.legend_pos = "bottom"},
willingness-to-pay threshold(s) \code{.wtp_threshold = c(20000, 30000)},
show WTP threshold(s) lines \code{.show_wtp = TRUE},
show WTP threshold(s) labels \code{.label_wtp = TRUE},
zoom to min/max values \code{.zoom = FALSE}, and
zoom to supplied coordinates values \code{.zoom_cords = NULL}.}
}
\value{
An object of class ggplot.
}
\description{
Plot Expected Net Monetary Benefit (eNMB)
}
\examples{
\dontrun{
library(ShinyPSA)

PSA_summary <- summarise_PSA_(
  .effs = as_tibble(ShinyPSA::Smoking_PSA$e),
  .costs = as_tibble(ShinyPSA::Smoking_PSA$c),
  .interventions = ShinyPSA::Smoking_PSA$treats)

p <- plot_eNMB_(PSA_summary,
               .legend_pos = NULL,
               .wtp_threshold = c(2000, 10000, 20000, 25000),
               .show_wtp = TRUE,
               .label_wtp = FALSE,
               .zoom = FALSE,
               .zoom_cords = NULL)

p
}

}
