% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_EVPI_.R
\name{plot_EVPI_}
\alias{plot_EVPI_}
\title{Plot the Expected Value of Perfect Information (EVPI)}
\usage{
plot_EVPI_(.PSA_data, ...)
}
\arguments{
\item{.PSA_data}{A list of class shinyPSA that contains summary PSA
results.}

\item{...}{Additional arguments that include:
legend position \code{.legend_pos = "bottom"},
willingness-to-pay threshold(s) \code{.wtp_threshold = c(20000, 30000)},
show WTP threshold(s) lines \code{.show_wtp = TRUE},
show WTP threshold(s) labels \code{.label_wtp = TRUE},
plot individual EVPI \code{.individual_evpi = TRUE},
time horizon to estimate population EVPI \code{.time_horion = 5},
discount rate to estimate population EVPI \code{.discount_rate = 0.035},
population size for population EVPI \code{.population = 15000},
zoom to min/max values \code{.zoom = FALSE}, and
zoom to supplied coordinates values \code{.zoom_cords = NULL}.}
}
\value{
An object of class ggplot.
}
\description{
Plot the Expected Value of Perfect Information (EVPI)
}
\examples{
\dontrun{
library(ShinyPSA)

PSA_summary <- summarise_PSA_(
  .effs = as_tibble(ShinyPSA::Smoking_PSA$e),
  .costs = as_tibble(ShinyPSA::Smoking_PSA$c),
  .interventions = ShinyPSA::Smoking_PSA$treats)

p <- plot_EVPI_(PSA_summary,
               .legend_pos = NULL,
               .wtp_threshold = c(2000, 10000, 20000, 25000),
               .show_wtp = TRUE,
               .label_wtp = FALSE,
               .individual_evpi = FALSE,
               .time_horion = 1,
               .discount_rate = 0.035,
               .population = 15000,
               .zoom = FALSE,
               .zoom_cords = NULL,
               .show_title = TRUE)

p
}

}
