% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_EVPPI_.R
\name{plot_EVPPI_}
\alias{plot_EVPPI_}
\title{Plot the Expected Value of Perfect Partial Information (EVPPI) results.}
\usage{
plot_EVPPI_(
  EVPPI_res,
  .show_percent = TRUE,
  .min_percent = 1,
  .params_num = NULL
)
}
\arguments{
\item{EVPPI_res}{A list containing EVPPI results.}

\item{.show_percent}{Bolean for whether to show the percentage of overall
EVPI on the bars.}

\item{.min_percent}{Only parameters with percentage of overall EVPI equal to
or higher than .min_percent will make it to the plot.}

\item{.params_num}{The number of parameters to show in the bar plot.}
}
\value{
An Object of class ggplot2.
}
\description{
Plot the Expected Value of Perfect Partial Information (EVPPI) results.
}
\examples{
\dontrun{
library(ShinyPSA)

# Summarise PSA results:
PSA_summary = summarise_PSA_(
  .effs = ShinyPSA::Brennan_1K_PSA$e,
  .costs = ShinyPSA::Brennan_1K_PSA$c,
  .params = ShinyPSA::Brennan_1K_PSA$p,
  .interventions = ShinyPSA::Brennan_1K_PSA$treats)

# Estimate EVPPI:
EVPPI_ind_res <- compute_EVPPIs_(
  .PSA_data = PSA_summary,
  .MAICER_ = 30000)

# Plot EVPPI results:
plot_EVPPI_(
  EVPPI_res = EVPPI_ind_res,
  .params_num = 15,
  .min_percent = NULL)
}
}
