% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_R6_elements.R
\name{prettyCheckboxGroup}
\alias{prettyCheckboxGroup}
\title{R6 Class employing shinyWidgets::prettyCheckboxGroup}
\format{
An \link[R6:R6Class]{R6::R6Class} object.
}
\description{
An instance of this class is used in similar way to
shinyWidgets::prettyCheckboxGroup
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-prettyCheckboxGroup-new}{\code{prettyCheckboxGroup$new()}}
\item \href{#method-prettyCheckboxGroup-ui_input}{\code{prettyCheckboxGroup$ui_input()}}
\item \href{#method-prettyCheckboxGroup-ui_output}{\code{prettyCheckboxGroup$ui_output()}}
\item \href{#method-prettyCheckboxGroup-ui_render_output}{\code{prettyCheckboxGroup$ui_render_output()}}
\item \href{#method-prettyCheckboxGroup-server}{\code{prettyCheckboxGroup$server()}}
\item \href{#method-prettyCheckboxGroup-get_uiInId}{\code{prettyCheckboxGroup$get_uiInId()}}
\item \href{#method-prettyCheckboxGroup-get_uiOtId}{\code{prettyCheckboxGroup$get_uiOtId()}}
\item \href{#method-prettyCheckboxGroup-get_uiRdOtId}{\code{prettyCheckboxGroup$get_uiRdOtId()}}
\item \href{#method-prettyCheckboxGroup-clone}{\code{prettyCheckboxGroup$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prettyCheckboxGroup-new"></a>}}
\if{latex}{\out{\hypertarget{method-prettyCheckboxGroup-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prettyCheckboxGroup$new(.label_)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prettyCheckboxGroup-ui_input"></a>}}
\if{latex}{\out{\hypertarget{method-prettyCheckboxGroup-ui_input}{}}}
\subsection{Method \code{ui_input()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prettyCheckboxGroup$ui_input(
  .tag_ = "UI_input",
  .choices_,
  .icon_ = "check",
  .bigger_ = TRUE,
  .animation_ = "tada",
  .status_ = "success",
  .class_ = "flex-fill text-right"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prettyCheckboxGroup-ui_output"></a>}}
\if{latex}{\out{\hypertarget{method-prettyCheckboxGroup-ui_output}{}}}
\subsection{Method \code{ui_output()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prettyCheckboxGroup$ui_output(.tag_ = "UI_output")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prettyCheckboxGroup-ui_render_output"></a>}}
\if{latex}{\out{\hypertarget{method-prettyCheckboxGroup-ui_render_output}{}}}
\subsection{Method \code{ui_render_output()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prettyCheckboxGroup$ui_render_output(.tag_ = "UI_render_output")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prettyCheckboxGroup-server"></a>}}
\if{latex}{\out{\hypertarget{method-prettyCheckboxGroup-server}{}}}
\subsection{Method \code{server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prettyCheckboxGroup$server(input, output, session)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prettyCheckboxGroup-get_uiInId"></a>}}
\if{latex}{\out{\hypertarget{method-prettyCheckboxGroup-get_uiInId}{}}}
\subsection{Method \code{get_uiInId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prettyCheckboxGroup$get_uiInId()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prettyCheckboxGroup-get_uiOtId"></a>}}
\if{latex}{\out{\hypertarget{method-prettyCheckboxGroup-get_uiOtId}{}}}
\subsection{Method \code{get_uiOtId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prettyCheckboxGroup$get_uiOtId()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prettyCheckboxGroup-get_uiRdOtId"></a>}}
\if{latex}{\out{\hypertarget{method-prettyCheckboxGroup-get_uiRdOtId}{}}}
\subsection{Method \code{get_uiRdOtId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prettyCheckboxGroup$get_uiRdOtId()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prettyCheckboxGroup-clone"></a>}}
\if{latex}{\out{\hypertarget{method-prettyCheckboxGroup-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{prettyCheckboxGroup$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
