% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{run_shiny_R6_App}
\alias{run_shiny_R6_App}
\title{Run the demo Shiny app.}
\usage{
run_shiny_R6_App(example_app = "ShinyPSA")
}
\arguments{
\item{example_app}{The example shiny app to run.}
}
\value{
Runs the example shiny app
}
\description{
Run the demo Shiny app.
}
\examples{
\dontrun{
run_shiny_R6_App()
}
}
