% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_R6_app.R
\name{ShinyPSA_R6_App}
\alias{ShinyPSA_R6_App}
\title{R6 Class representing a Shiny app.}
\format{
An \link[R6:R6Class]{R6::R6Class} object.
}
\description{
An instance of this class is a fully fledged shiny app.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ShinyPSA_R6_App-new}{\code{ShinyPSA_R6_App$new()}}
\item \href{#method-ShinyPSA_R6_App-ui}{\code{ShinyPSA_R6_App$ui()}}
\item \href{#method-ShinyPSA_R6_App-waiter_dev}{\code{ShinyPSA_R6_App$waiter_dev()}}
\item \href{#method-ShinyPSA_R6_App-server}{\code{ShinyPSA_R6_App$server()}}
\item \href{#method-ShinyPSA_R6_App-clone}{\code{ShinyPSA_R6_App$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShinyPSA_R6_App-new"></a>}}
\if{latex}{\out{\hypertarget{method-ShinyPSA_R6_App-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyPSA_R6_App$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShinyPSA_R6_App-ui"></a>}}
\if{latex}{\out{\hypertarget{method-ShinyPSA_R6_App-ui}{}}}
\subsection{Method \code{ui()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyPSA_R6_App$ui()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShinyPSA_R6_App-waiter_dev"></a>}}
\if{latex}{\out{\hypertarget{method-ShinyPSA_R6_App-waiter_dev}{}}}
\subsection{Method \code{waiter_dev()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyPSA_R6_App$waiter_dev(.info_)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShinyPSA_R6_App-server"></a>}}
\if{latex}{\out{\hypertarget{method-ShinyPSA_R6_App-server}{}}}
\subsection{Method \code{server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyPSA_R6_App$server(input, output, session)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShinyPSA_R6_App-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ShinyPSA_R6_App-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyPSA_R6_App$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
