% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{stats_generator_}
\alias{stats_generator_}
\title{Generate stats}
\usage{
stats_generator_(.effs, .costs, .interventions, .units_ = "£")
}
\arguments{
\item{.effs}{A matrix containing the \code{effects} from PSA. Number of
\code{columns} is equal to the interventions while the number of
\code{rows} is equal to the number of PSA simulations to be summarised.}

\item{.costs}{A matrix containing the \code{costs} from PSA. Number of
\code{columns} is equal to the interventions while the number of
\code{rows} is equal to the number of PSA simulations to be summarised.}

\item{.interventions}{A vector containing the names of all
interventions. If not provided or less names than needed is provided,
the function will generate generic names, for example
\code{intervention 1}.}

\item{.units_}{A character, the units to associate with the
monitory values in the summary table. Default is sterling pounds
(GBP) \code{"\u00A3"}.}
}
\description{
Generate stats
}
