% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_PSA_.R
\name{summarise_PSA_}
\alias{summarise_PSA_}
\title{Summarise PSA outputs and report results}
\usage{
summarise_PSA_(
  .effs,
  .costs,
  .interventions = NULL,
  .ref = NULL,
  .Kmax = 1e+05,
  .wtp = NULL,
  .plot = FALSE
)
}
\arguments{
\item{.effs}{A matrix containing the \code{effects} from PSA. Number of
\code{columns} is equal to the interventions while the number of
\code{rows} is equal to the number of PSA simulations to be summarised.}

\item{.costs}{A matrix containing the \code{costs} from PSA. Number of
\code{columns} is equal to the interventions while the number of
\code{rows} is equal to the number of PSA simulations to be summarised.}

\item{.interventions}{A vector containing the names of all
interventions. If not provided or less names than needed is provided,
the function will generate generic names, for example
\code{intervention 1}.}

\item{.ref}{An integer indicating the index of the reference
intervention. This parameter is ignored if more than two
\code{interventions} are under analysis.}

\item{.Kmax}{The maximum willingness-to-pay threshold to use in the
analysis. This parameter is ignored if \code{wtp} is provided.}

\item{.wtp}{A vector of numerical values declaring the
willingness-to-pay (WTP) values to use in the analysis. If \code{NULL}
(default) a range of WTP values (up to \code{.Kmax} will be used.}

\item{.plot}{A boolean, FALSE (default), for whether to generate plots.}
}
\value{
A list of class \code{psa} with \code{24} elements.
}
\description{
Summarise PSA outputs and report results
}
\examples{
\dontrun{}
}
