% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_R6_elements.R
\name{textInput}
\alias{textInput}
\title{R6 Class employing shiny::textInput}
\format{
An \link[R6:R6Class]{R6::R6Class} object.
}
\description{
An instance of this class is used in similar way to shiny::textInput
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-textInput-new}{\code{textInput$new()}}
\item \href{#method-textInput-ui_input}{\code{textInput$ui_input()}}
\item \href{#method-textInput-ui_output}{\code{textInput$ui_output()}}
\item \href{#method-textInput-ui_render_output}{\code{textInput$ui_render_output()}}
\item \href{#method-textInput-server}{\code{textInput$server()}}
\item \href{#method-textInput-get_uiInId}{\code{textInput$get_uiInId()}}
\item \href{#method-textInput-get_uiOtId}{\code{textInput$get_uiOtId()}}
\item \href{#method-textInput-get_uiRdOtId}{\code{textInput$get_uiRdOtId()}}
\item \href{#method-textInput-clone}{\code{textInput$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-textInput-new"></a>}}
\if{latex}{\out{\hypertarget{method-textInput-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{textInput$new(.label_)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-textInput-ui_input"></a>}}
\if{latex}{\out{\hypertarget{method-textInput-ui_input}{}}}
\subsection{Method \code{ui_input()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{textInput$ui_input(
  .tag_ = "UI_input",
  .width_ = "100\%",
  .class_ = "d-flex align-items-center",
  .value_ = "500, 20000, 30000, 50000"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-textInput-ui_output"></a>}}
\if{latex}{\out{\hypertarget{method-textInput-ui_output}{}}}
\subsection{Method \code{ui_output()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{textInput$ui_output(.tag_ = "UI_output")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-textInput-ui_render_output"></a>}}
\if{latex}{\out{\hypertarget{method-textInput-ui_render_output}{}}}
\subsection{Method \code{ui_render_output()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{textInput$ui_render_output(.tag_ = "UI_render_output")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-textInput-server"></a>}}
\if{latex}{\out{\hypertarget{method-textInput-server}{}}}
\subsection{Method \code{server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{textInput$server(input, output, session, .min_, .max_)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-textInput-get_uiInId"></a>}}
\if{latex}{\out{\hypertarget{method-textInput-get_uiInId}{}}}
\subsection{Method \code{get_uiInId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{textInput$get_uiInId()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-textInput-get_uiOtId"></a>}}
\if{latex}{\out{\hypertarget{method-textInput-get_uiOtId}{}}}
\subsection{Method \code{get_uiOtId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{textInput$get_uiOtId()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-textInput-get_uiRdOtId"></a>}}
\if{latex}{\out{\hypertarget{method-textInput-get_uiRdOtId}{}}}
\subsection{Method \code{get_uiRdOtId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{textInput$get_uiRdOtId()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-textInput-clone"></a>}}
\if{latex}{\out{\hypertarget{method-textInput-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{textInput$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
